Refactor helpers into separate modules
